refactor schema loader to reduce code duplication
